charge_point_data_process/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

pandas
polars 
pyarrow
numpy
psycopg2
azure-storage-blob
aiohttp
geopy
scipy